requirements.txt
pandas>=1.0
requests
tqdm
dask[complete]
beautifulsoup4
psycopg2-binary
sqlalchemy
stringcase
d6tstack[psql]
